requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
chromadb==0.5.5
sentence-transformers==3.1.0
pdfplumber==0.11.4
PyMuPDF==1.24.10
pydantic==2.9.1
requests==2.32.3
python-dotenv==1.0.1